Add a "/V1/metrics/orders" REST API endpoint
Using the regular "/V1/orders" REST API endpoint is a performance
bottleneck when Magento contains a considerable amount of orders.

Magento (regardless of the \"fields\" passed to the API request)
will query for a lot more than you asked for.

Returning the data we need for all orders for the exporter to then
aggregate also results in a unnecessary large response size.